feat: add jotai and track auth state
