+++ app/_layout.tsx
@@ -1,20 +1,21 @@
-import "react-native-url-polyfill/auto";
-import { supabase } from "@/database/supabase.client";
 import { GluestackUIProvider } from "@/components/ui/gluestack-ui-provider";
+import { supabase } from "@/database/supabase.client";
 import "@/global.css";
 import { useColorScheme } from "@/hooks/useColorScheme";
+import { sessionAtom } from "@/state/auth";
 import {
   DarkTheme,
   DefaultTheme,
   ThemeProvider,
 } from "@react-navigation/native";
-import { Session } from "@supabase/supabase-js";
 import { useFonts } from "expo-font";
-import { Stack } from "expo-router";
+import { useSegments, Stack, usePathname, useRouter } from "expo-router";
 import * as SplashScreen from "expo-splash-screen";
 import { StatusBar } from "expo-status-bar";
-import { useEffect, useState } from "react";
+import { useAtomValue, useSetAtom } from "jotai";
+import { useEffect } from "react";
 import "react-native-reanimated";
+import "react-native-url-polyfill/auto";
 import "../hooks/useLog"; // import useLog hook for global registration
 import "../util/log"; // relative path for reliable module resolution
 
@@ -22,9 +23,7 @@ import "../util/log"; // relative path for reliable module resolution
 SplashScreen.preventAutoHideAsync();
 
 export default function RootLayout() {
-  const [session, setSession] = useState<Session | null>(null);
-
-  useLog({ user: session?.user.email });
+  const setSession = useSetAtom(sessionAtom);
 
   useEffect(() => {
     supabase.auth.getSession().then(({ data: { session } }) => {
@@ -34,7 +33,7 @@ export default function RootLayout() {
     supabase.auth.onAuthStateChange((_event, session) => {
       setSession(session);
     });
-  }, []);
+  }, [setSession]);
 
   const colorScheme = useColorScheme();
 
@@ -55,14 +54,37 @@ export default function RootLayout() {
   return (
     <GluestackUIProvider mode={colorScheme === "dark" ? "dark" : "light"}>
       <ThemeProvider value={colorScheme === "dark" ? DarkTheme : DefaultTheme}>
-        <Stack>
-          <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+        <HasAuth>
+          <Stack>
+            <Stack.Screen
+              name="authenticated/(tabs)"
+              options={{ headerShown: false }}
+            />
 
-          <Stack.Screen name="+not-found" />
-        </Stack>
+            <Stack.Screen name="+not-found" />
+          </Stack>
 
-        <StatusBar style="auto" />
+          <StatusBar style="auto" />
+        </HasAuth>
       </ThemeProvider>
     </GluestackUIProvider>
   );
 }
+
+const HasAuth = ({ children }: { children: React.ReactNode }) => {
+  const session = useAtomValue(sessionAtom);
+
+  const segments = useSegments();
+
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!session && segments.at(0) !== "public") {
+      console.log("redirecting to sign-in");
+
+      return router.replace("/public/sign-in");
+    }
+  }, [session, segments, router]);
+
+  return <>{children}</>;
+};
