+++ app/authenticated/(tabs)/_layout.tsx
@@ -1,7 +1,6 @@
 import { Tabs } from "expo-router";
 import React from "react";
 import { Platform } from "react-native";
-
 import { HapticTab } from "@/components/HapticTab";
 import { IconSymbol } from "@/components/IconSymbol";
 import TabBarBackground from "@/components/TabBarBackground";
