+++ app/authenticated/(tabs)/explore.tsx
