+++ package.json
@@ -102,7 +102,8 @@
     "react-native-url-polyfill": "^2.0.0",
     "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.5",
-    "tailwindcss": "3.4.17"
+    "tailwindcss": "3.4.17",
+    "jotai": "^2.12.1"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
